feat: drain pod before delete node
