+++ pkg/apis/kosmos/v1alpha1/global_node_types.go
@@ -13,6 +13,7 @@ import (
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 // +kubebuilder:printcolumn:name="NODE_IP",type=string,JSONPath=`.spec.nodeIP`
 // +kubebuilder:printcolumn:name="STATE",type=string,JSONPath=`.spec.state`
+// +kubebuilder:printcolumn:name="VIRTUAL_CLUSTER",type=string,JSONPath=`.status.VirtualCluster`
 
 type GlobalNode struct {
 	metav1.TypeMeta   `json:",inline"`
